Rename util/files.py to util/file.py
